Split knowledge base: client-safe builtin module without firebase-admin
/knowledge-base page imported lib/knowledge-base which pulled firebase-admin
into the client bundle (child_process error). Built-in catalog and search
live in knowledge-base-builtin.ts; server Firestore helpers stay in
knowledge-base.ts.
